final project/requirements.txt
numpy
imageio
scipy
torch
torchvision
opencv-python
dlib
p-utils
Pillow